requirements.txt
flask
redis
python-dotenv
psycopg2-binary
